Rework operand counting and verify atoms
Operands and results counts may depend on the elements
currently on the various stacks. This change allows the
atoms to specify their stack requirements based on
that information.

Additionally, verification is now better handled by the
context, allowing the Verify atoms to specify their stack
usage in the normal case, while letting the context deal
with verification in a standard way when an atom
requires it.